Add confirmation message box when deleting custom zone layout

+++ src/modules/fancyzones/editor/FancyZonesEditor/MainWindow.xaml.cs
@@ -177,7 +177,13 @@ namespace FancyZonesEditor
                 SetSelectedItem();
             }
 
-            model.Delete();
+            string msgBoxName = "Delete Confirmation";
+            string msgBoxText = "Delete \"" + model.Name + "\" custom layout?";
+            MessageBoxResult confirmation = MessageBox.Show(msgBoxText, msgBoxName, MessageBoxButton.YesNo);
+            if (confirmation == MessageBoxResult.Yes)
+            {
+                model.Delete();
+            }
         }
 
         private void OnEdit(object sender, RoutedEventArgs e)
